Xspress fly scanning now fills in missing data points.
The fly-scan mechanism relies on camonitors to record the data. If a
given PV doesn't change from one frame to the next (e.g. the ROI is 0
both times), then no camonitor is sent. The xspress device now fills
in these missing data with the last recorded value for the given
signal.