Re-wrote the ion chamber scaler counts-to-volts calculation.
Previously, we assumed the clock frequency for channel 1 was the same
as the max frequency of the V-to-F source (e.g. V2F100), and that the
max V-to-F is 10V. This change removes that assumption, and the max
source frequency and max voltage of the V-to-F are now an iconfig
parameter (counts_per_volt_second).