Added a check to the undulator positioners to ensure that gap deadband is turned off.
If the gap deadband is on, then for moves with the deadband the ID
does not attempt to move, so the positioner times out. Chengjun Sun
and Mark Wolfman decided that this check was better than just turned
it off in the stage()/unstage() process because there may be
situations where the deadband is intended and this will avoid silent
unexpected behavior.